pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "team-reports"
version = "1.0.0"
description = "Automated team reporting from Jira and GitHub with weekly, quarterly, and performance analytics"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Chris Chase", email = "cochase@redhat.com"}
]
keywords = ["jira", "github", "reports", "analytics", "performance", "team-management"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "mcp>=0.9.0",
    "jira>=3.8.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "requests>=2.31.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.11.0",
]

[project.urls]
Homepage = "https://github.com/cmchase/team-reports"
Repository = "https://github.com/cmchase/team-reports"
Documentation = "https://github.com/cmchase/team-reports#readme"

[project.scripts]
team-reports = "team_reports.cli.main:cli"

[tool.setuptools]
packages = ["team_reports", "team_reports.reports", "team_reports.cli", "team_reports.utils"]

[tool.setuptools.package-data]
team_reports = ["py.typed"]
